oauth: validate clean refact
