+++ src/backend/src/auth/auth.controller.spec.ts
@@ -1,3 +1,4 @@
+import { HttpModule } from '@nestjs/axios';
 import { JwtService } from '@nestjs/jwt';
 import { Test, TestingModule } from '@nestjs/testing';
 import { DbService } from 'src/db/db.service';
@@ -11,6 +12,7 @@ describe('AuthController', () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [AuthController],
       providers: [AuthService, DbService, JwtService],
+      imports: [HttpModule],
     }).compile();
 
     controller = module.get<AuthController>(AuthController);
