+++ src/backend/src/auth/strategy/oauth2.strategy.ts
@@ -1,15 +1,14 @@
-import { HttpService } from '@nestjs/axios';
 import { Injectable, Logger } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy, StrategyOptions, VerifyCallback } from 'passport-oauth2';
-import { lastValueFrom } from 'rxjs';
+import { AuthService } from '../auth.service';
 
 @Injectable()
 export class OAuth2Strategy extends PassportStrategy(Strategy, 'oauth2') {
   private readonly logger = new Logger(OAuth2Strategy.name);
 
   // @doc http://www.passportjs.org/packages/passport-oauth2/
-  constructor(private readonly httpService: HttpService) {
+  constructor(private readonly authService: AuthService) {
     super(<StrategyOptions>{
       authorizationURL: 'https://api.intra.42.fr/oauth/authorize',
       tokenURL: 'https://api.intra.42.fr/oauth/token',
@@ -28,15 +27,7 @@ export class OAuth2Strategy extends PassportStrategy(Strategy, 'oauth2') {
     this.logger.debug(`${OAuth2Strategy.name}.${this.validate.name} called`);
     this.logger.debug(`accessToken: ${accessToken}`);
     this.logger.debug(`refreshToken: ${refreshToken}`);
-    const userObservable = this.httpService.get(
-      'https://api.intra.42.fr/v2/me',
-      {
-        headers: { Authorization: `Bearer ${accessToken}` },
-      },
-    );
-    const { data } = await lastValueFrom(userObservable); // lastValueFrom: convert observable to promise
-    // this.logger.debug(`data: ${JSON.stringify(data, null, 2)}`); // user format: https://api.intra.42.fr/apidoc/2.0/users/me.html
-    const user = { email: data.email, login: data.login };
+    const user = this.authService.oauthFindOrCreate(accessToken);
     verified(null, user, { accessToken, refreshToken, profile });
   }
 }
